% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R, R/zzz.R
\name{correlation_heatmaps}
\alias{correlation_heatmaps}
\alias{correlation.heatmaps}
\title{Heatmaps of top correlated genes}
\usage{
correlation_heatmaps(
  seurat.object,
  correlation.df,
  up.n = 20,
  dn.n = 20,
  goi = "NKG7",
  output.dir = "results/figures/Correlation_analysis/"
)

correlation.heatmaps(
  seurat.object,
  correlation.df,
  up.n = 20,
  dn.n = 20,
  goi = "NKG7",
  output.dir = "results/figures/Correlation_analysis/"
)
}
\arguments{
\item{seurat.object}{A Seurat object (idents/assay pre-set, ideally downsampled).}

\item{correlation.df}{Output of \code{\link[=sc_correlation]{sc_correlation()}}.}

\item{up.n, dn.n}{Number of up / down correlated genes to show.}

\item{goi}{Gene of interest.}

\item{output.dir}{Directory to write the PDFs into.}
}
\value{
Invisibly \code{NULL}; called for the PDFs it writes.
}
\description{
Draws Seurat heatmaps of the most positively and negatively correlated genes
from \code{\link[=sc_correlation]{sc_correlation()}}, using the batlow palette.
}
\examples{
\dontrun{
correlation_heatmaps(small_seurat, cor_df, up.n = 20, goi = "NKG7")
}
}
