% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R, R/zzz.R
\name{get_batlow}
\alias{get_batlow}
\alias{Get.batlow}
\title{Batlow colour palette}
\usage{
get_batlow(n = 100, rev = FALSE)

Get.batlow(n = 100, rev = FALSE)
}
\arguments{
\item{n}{Number of colours to return (>= 1).}

\item{rev}{If \code{TRUE}, reverse the palette order.}
}
\value{
A character vector of hex colours.
}
\description{
Returns \code{n} colours from the perceptually-uniform \pkg{scico} "batlow" palette.
}
\examples{
get_batlow(5)
get_batlow(5, rev = TRUE)
}
