% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{plot_correlation}
\alias{plot_correlation}
\title{Plot a correlation ranking}
\usage{
plot_correlation(
  cor.df,
  output.dir = "output/figures/",
  goi = "NKG7",
  top.n.val = 30
)
}
\arguments{
\item{cor.df}{Output of \code{\link[=sc_correlation]{sc_correlation()}}.}

\item{output.dir}{Directory to write the PDFs into.}

\item{goi}{Gene(s) of interest to highlight.}

\item{top.n.val}{Number of top-correlated genes to inset.}
}
\value{
Invisibly \code{NULL}; called for the PDFs it writes.
}
\description{
Line plot of ranked correlation values from \code{\link[=sc_correlation]{sc_correlation()}}, highlighting a
gene of interest and the top hits.
}
\examples{
\dontrun{
plot_correlation(cor_df, goi = "NKG7", top.n.val = 30)
}
}
