% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R, R/colours.R
\name{Get.batlow}
\alias{Get.batlow}
\alias{get_batlow}
\title{Batlow colour palette}
\usage{
Get.batlow(n = 100)

get_batlow(n = 100)
}
\arguments{
\item{n}{Number of colours to return.}
}
\value{
A character vector of hex colours.
}
\description{
Returns \code{n} colours from the perceptually-uniform \pkg{scico} "batlow" palette.
}
\examples{
get_batlow(5)
}
